seperated udp logging from others
