+++ tests/tests_base/test_logging/.gitignore
@@ -1,0 +1,1 @@
+logs/
