+++ tests/tests_base/test_logging_udp/.gitignore
@@ -1,0 +1,1 @@
+logs/
